+++ apps/diffusion/diffusion.cpp
@@ -24,12 +24,6 @@
 
 #include "colormanip.h"
 
-#include "../../config.h"
-
-#ifdef HAVE_SOLVER_WRAPPERS
-    #include "agmg/agmg.hpp"
-#endif
-
 #include "loaders/loader.hpp"
 #include "hho/hho.hpp"
