+++ contrib/jburkardt/CMakeLists.txt
@@ -3,8 +3,4 @@ set(quadratures_source
     ${CMAKE_CURRENT_SOURCE_DIR}/jb_common.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/simplex_gm_rule.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tetrahedron_arbq_rule.cpp
-    )
-
-add_library(quadratures SHARED ${quadratures_source})
-set(LINK_LIBS "${LINK_LIBS}" quadratures PARENT_SCOPE)
-install(TARGETS quadratures LIBRARY DESTINATION lib)
+    PARENT_SCOPE)
