+++ core/solvers/agmg.hpp
@@ -44,9 +44,6 @@
 
 #pragma once
 
-
-
-
 #include <Eigen/Dense>
 #include <Eigen/Sparse>
